storage/certificate-service/requirements.txt
psycopg2-binary>=2.9.9
reportlab>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
